Update Get subscription after changing lights
The previous code did not work correctly because the local
copies were in fact references to the light objects in huejay.
Therefore the publish calls to the "get" topics were never updated
if a change came from mqtt-hue.

This commit updates the get topic for all lights once its saved.