Use HueJay Timeout Value
Previously a timeout of 50 was used, resulting in a 50ms timeout,
which is quite low and resulting in mqtt-hue hammering
Hue bridges that cannot respond to requests within that interval.
HueJays default timeout is 15000 (15s).
The added benefit of adjusting this is a more stable connection to
the Hue bridge as well as less CPU Usage on platforms such as Raspberry
Pi.